Begin to implement CPU flags. Move instructions into their own module
