+++ src/cpu/instructions/misc.rs
